backend/requirements.txt
# Backend Dependencies for CIMEIKA

# Web Frameworks
Flask==3.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Redis
redis==5.0.1
hiredis==2.2.3

# Async tasks
celery==5.3.4

# AI Integration
openai==1.3.7
anthropic==0.7.7

# HTTP & API
requests==2.31.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
PyYAML==6.0.1

# CORS
flask-cors==4.0.0

# Monitoring
sentry-sdk==1.39.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
